Move editor helper methods to a specific file
